backend/requirements.txt
azure-functions==1.20.0
pydantic==2.8.2
python-docx==1.1.2
